Add Initial version of Analysis Module. - Add options to select which configuration to run (Simple, Simple + Statistical or Simple, Statistical + Analysis - Minor stability and performance improvements.

+++ src/org/nerif/ml/DecisionNode.java
@@ -5,7 +5,7 @@ import java.util.HashMap;
 public class DecisionNode {
 	public int col;
 	public String value;
-	public HashMap<String, Integer> results;
+	public HashMap<String, Long> results;
 	public DecisionNode tb;
 	public DecisionNode fb;
